backend/pyproject.toml
[project]
name = "blueprint4fastapi-backend"
version = "0.1.0"
description = "FastAPI auth template for agent-driven customization"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "redis[hiredis]>=5.0.0",
    "fakeredis>=2.23.0",
]

[tool.uv]
package = false

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
